REV (2)  main 구조체내의 CardDeck 셔플, 뽑기 로직을 별도 클래스로 분리
